netguard/requirements.txt
scapy>=2.5.0
pyshark>=0.6
scikit-learn>=1.3
xgboost>=2.0
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
colorama>=0.4
jinja2>=3.1
requests>=2.31
